+++ CHANGES.md
@@ -46,6 +46,8 @@
 * Add scene qualities WEB.h264 to SDTV, 720p.WEB.h264 to WEB DL 720p, and 1080p.WEB.h264 to WEB DL 1080p
 * Change improve handling when provider PiSexy is missing expected data
 * Change show list second level sort criteria
+* Change adding show processing to be highest priority
+* Use timezones to check unaired status during show update/adding
 
 
 ### 0.11.10 (2016-03-17 19:00:00 UTC)